api/pyproject.toml
[project]
name = "smia-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.4.0",
    "fastapi>=0.129.0",
    "firecrawl-py>=4.14.1",
    "httpx>=0.28.1",
    "langfuse>=3.14.1",
    "mangum>=0.21.0",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.59.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "supabase>=2.28.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
